+++ misc/_gg.zsh
@@ -87,7 +87,7 @@ case "${words[2]}" in
       ':command:' \
       {-d,-delete}'[delete the given branch]' \
       {-f,-force}'[force]' \
-      {-p,-pattern}'=[regexp of branches to list]' \
+      '*'{-p,-pattern}'=[regexp of branches to list]' \
       '-r=[revision]:rev:named_revs' \
       '-sort=[sort order for listing]:order:(name -name date -date)' \
       '*:name:branches'