+++ misc/_gg.zsh
@@ -206,6 +206,7 @@ case "${words[2]}" in
     _arguments -S : \
       ':command:' \
       '-r=[remote reference intended to be pulled]:remote ref:branches' \
+      '*'{-p,-pattern}'=[regexp of branch or tag names to pull]' \
       '-force-tags[update any tags pulled]' \
       '-u[update to new head if new descendants were pulled]' \
       ':source:remotes'